requirements.txt
twilio==9.0.5
Flask==3.0.3
flask-sock==0.7.0
ngrok==1.3.0
python-dotenv==1.0.1
deepgram-sdk==3.2.7
openai==1.30.1
numpy==1.26.4
requests==2.32.2
PyAudio==0.2.14
Flask-Cors==4.0.1









